requirements.txt
requests==2.26.0
Django==4.2
djangorestframework==3.14.0
haversine==2.8.0
Pillow==8.4.0
PyMySQL==1.0.3


